requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
certifi==2022.5.18.1
    # via requests
charset-normalizer==2.0.12
    # via requests
idna==3.3
    # via requests
lxml==4.8.0
    # via webdavclient3
python-dateutil==2.8.2
    # via webdavclient3
python-dotenv==0.20.0
    # via -r requirements.in
requests==2.27.1
    # via
    #   -r requirements.in
    #   webdavclient3
six==1.16.0
    # via python-dateutil
urllib3==1.26.9
    # via requests
webdavclient3==3.14.6
    # via -r requirements.in
